==> 2020/03_State_popup.Rmd <==
---
title: "Popups"
author: "Diana Villasana Ocampo"
output:
   html_document:
      highlight: tango
      theme: flatly
      toc: yes
      toc_depth: 2
      toc_float:
        collapsed: yes
---

```{r, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, message = FALSE, warning = FALSE, cache = FALSE, cache.lazy = FALSE, 
                         eval = FALSE, class.source = "fold-show")
knitr::opts_knit$set(root.dir = rprojroot::find_rstudio_root_file())
options(digits = 2, encoding = "UTF8")
```   
 
```{r, echo=FALSE}
rm(list = ls())
```

```{r, echo=FALSE}
setwd(here::here())
```

**Librerías**
```{r}
require(dplyr)
require(forcats)
require(tibble)
require(sf)
require(sp)
require(stringr)
require(dplyr)
require(leaflet)
require(RColorBrewer)
```
# Base de datos   

```{r}
load(paste0(here::here(), "/2020/Output/IMM_2020.RData"))
```

# Shapefiles 

```{r}
## Shapefiles
shape_municipios_1 <- geojsonio::geojson_read(paste0(here::here(), "/2020/Output/Shapes/municipios_1_json.geojson"), what = "sp") 
shape_municipios_2 <- geojsonio::geojson_read(paste0(here::here(), "/2020/Output/Shapes/municipios_2_json.geojson"), what = "sp") 
shape_municipios_3 <- geojsonio::geojson_read(paste0(here::here(), "/2020/Output/Shapes/municipios_3_json.geojson"), what = "sp") 
shape_municipios_4 <- geojsonio::geojson_read(paste0(here::here(), "/2020/Output/Shapes/municipios_4_json.geojson"), what = "sp") 
shape_municipios_5 <- geojsonio::geojson_read(paste0(here::here(), "/2020/Output/Shapes/municipios_5_json.geojson"), what = "sp") 
```


# Paleta de colores 

```{r}
paleta <- c("#13322B", #Muy alto
            "#086953", #Alto 
            "#4C9562", #Medio
            "#D1BB9E", #Bajo
            "#C3BA6D") # Muy bajo


mypalette <- leaflet::colorFactor(palette = paleta, 
                                   domain = forcats::fct_relevel(IMM_2020$GM_2020, c("Muy alto", "Alto", "Medio", "Bajo", "Muy bajo")), 
                                   na.color = "transparent")
```


## Popup `Muy alto`

```{r}
## Paleta de colores                         
fillColor <- paleta[1]

df_municipio <- IMM_2020 %>% 
                 mutate(across(.cols = c(5:15,17), as.numeric)) %>%
                  filter(GM_2020 %in% "Muy alto")

## Etiquetas del mapa a nivel municipal 
state_popup_municipio <- paste('<font size="3"; font-family: "Century Gothic"><strong>',"Estado:",'</strong>','<strong>', df_municipio$NOM_ENT,'</strong></font>','<br/>',
                               '<font size="2.5"; font-family: "Century Gothic"><strong>',"Municipio:",'</strong>','<strong>', df_municipio$NOM_MUN,'</strong></font>','<br/>',
                               '<font size="2"; font-family: "Century Gothic"><strong>',"Población Total:",'</strong></font>','<font size="2"; font-family: "Century Gothic"; color=', fillColor,'><strong>', formatC(df_municipio$POB_TOT, format = "f", big.mark = " ", digits = 0),'</strong></font>','<br/>',
                               '<font size="2"; font-family: "Century Gothic"><strong>',"Grado de marginación:",'</strong></font>','<font size="2"; font-family: "Century Gothic"; color=', fillColor,'><strong>', df_municipio$GM_2020,'</strong></font>','<br/>',
                               '<font size="2"; font-family: "Century Gothic"><strong>',"Índice de marginación:",'</strong></font>','<font size="2"; font-family: "Century Gothic"; color=', fillColor,'><strong>', format(round(df_municipio$IM_2020, 2), nsmall = 1),'</strong></font>','<br/>',
                               '<font size="2"; font-family: "Century Gothic"><strong>',"Índice de marginación normalizado:",'</strong></font>','<font size="2"; font-family: "Century Gothic"; color=', fillColor,'><strong>', format(round(df_municipio$IMN_2020, 2), nsmall = 1),'</strong></font>','<br/>',
                               '<font size="1.5"; font-family: "Century Gothic"><strong>',"% Población de 15 años y más analfabeta:",'</strong>', round(df_municipio$ANALF, 2),'</font>','<br/>',
                               '<font size="1.5"; font-family: "Century Gothic"><strong>',"% Población de 15 años o más sin educación básica:",'</strong>', round(df_municipio$SBASC, 2),'</font>','<br/>',
                               '<font size="1.5"; font-family: "Century Gothic"><strong>',"% Ocupantes en viviendas paticulares sin drenaje ni excusado:",'</strong>', round(df_municipio$OVSDE, 2),'<br/>', 
                               '<font size="1.5"; font-family: "Century Gothic"><strong>',"% Ocupantes en viviendas paticulares sin energía eléctrica:",'</strong>', round(df_municipio$OVSEE, 2),'</font>','<br/>', 
                               '<font size="1.5"; font-family: "Century Gothic"><strong>',"% Ocupantes en viviendas paticulares sin agua entubada:",'</strong>', round(df_municipio$OVSAE, 2),'</font>','<br/>',
                               '<font size="1.5"; font-family: "Century Gothic"><strong>',"% Ocupantes en viviendas paticulares con piso de tierra:",'</strong>', round(df_municipio$OVPT, 2),'</font>','<br/>',
                               '<font size="1.5"; font-family: "Century Gothic"><strong>',"% Viviendas paticulares con algún nivel de hacinamiento:",'</strong>', round(df_municipio$VHAC, 2),'</font>','<br/>',
                               '<font size="1.5"; font-family: "Century Gothic"><strong>',"% Población en localidades con menos de \n 5 000 habitantes:",'</strong>', round(df_municipio$PL.5000, 2),'</font>','<br/>',
                               '<font size="1.5"; font-family: "Century Gothic"><strong>',"% Población ocupada con ingresos menores a 2SM:",'</strong>', round(df_municipio$PO2SM, 2),'</font>'
) %>% 
  lapply(htmltools::HTML)

saveRDS(state_popup_municipio, paste0(here::here(), "/2020/Output/PopUp/state_popup_municipio1.RDS"))
```


## Popup `Alto`

```{r}
## Paleta de colores                         
fillColor <- paleta[2]

df_municipio <- IMM_2020 %>% 
                 mutate(across(.cols = c(5:15,17), as.numeric)) %>%
                  filter(GM_2020 %in% "Alto")

## Etiquetas del mapa a nivel municipal 
state_popup_municipio <- paste('<font size="3"; font-family: "Century Gothic"><strong>',"Estado:",'</strong>','<strong>', df_municipio$NOM_ENT,'</strong></font>','<br/>',
                               '<font size="2.5"; font-family: "Century Gothic"><strong>',"Municipio:",'</strong>','<strong>', df_municipio$NOM_MUN,'</strong></font>','<br/>',
                               '<font size="2"; font-family: "Century Gothic"><strong>',"Población Total:",'</strong></font>','<font size="2"; font-family: "Century Gothic"; color=', fillColor,'><strong>', formatC(df_municipio$POB_TOT, format = "f", big.mark = " ", digits = 0),'</strong></font>','<br/>',
                               '<font size="2"; font-family: "Century Gothic"><strong>',"Grado de marginación:",'</strong></font>','<font size="2"; font-family: "Century Gothic"; color=', fillColor,'><strong>', df_municipio$GM_2020,'</strong></font>','<br/>',
                               '<font size="2"; font-family: "Century Gothic"><strong>',"Índice de marginación:",'</strong></font>','<font size="2"; font-family: "Century Gothic"; color=', fillColor,'><strong>', format(round(df_municipio$IM_2020, 2), nsmall = 1),'</strong></font>','<br/>',
                               '<font size="2"; font-family: "Century Gothic"><strong>',"Índice de marginación normalizado:",'</strong></font>','<font size="2"; font-family: "Century Gothic"; color=', fillColor,'><strong>', format(round(df_municipio$IMN_2020, 2), nsmall = 1),'</strong></font>','<br/>',
                               '<font size="1.5"; font-family: "Century Gothic"><strong>',"% Población de 15 años y más analfabeta:",'</strong>', round(df_municipio$ANALF, 2),'</font>','<br/>',
                               '<font size="1.5"; font-family: "Century Gothic"><strong>',"% Población de 15 años o más sin educación básica:",'</strong>', round(df_municipio$SBASC, 2),'</font>','<br/>',
                               '<font size="1.5"; font-family: "Century Gothic"><strong>',"% Ocupantes en viviendas paticulares sin drenaje ni excusado:",'</strong>', round(df_municipio$OVSDE, 2),'<br/>', 
                               '<font size="1.5"; font-family: "Century Gothic"><strong>',"% Ocupantes en viviendas paticulares sin energía eléctrica:",'</strong>', round(df_municipio$OVSEE, 2),'</font>','<br/>', 
                               '<font size="1.5"; font-family: "Century Gothic"><strong>',"% Ocupantes en viviendas paticulares sin agua entubada:",'</strong>', round(df_municipio$OVSAE, 2),'</font>','<br/>',
                               '<font size="1.5"; font-family: "Century Gothic"><strong>',"% Ocupantes en viviendas paticulares con piso de tierra:",'</strong>', round(df_municipio$OVPT, 2),'</font>','<br/>',
                               '<font size="1.5"; font-family: "Century Gothic"><strong>',"% Viviendas paticulares con algún nivel de hacinamiento:",'</strong>', round(df_municipio$VHAC, 2),'</font>','<br/>',
                               '<font size="1.5"; font-family: "Century Gothic"><strong>',"% Población en localidades con menos de \n 5 000 habitantes:",'</strong>', round(df_municipio$PL.5000, 2),'</font>','<br/>',
                               '<font size="1.5"; font-family: "Century Gothic"><strong>',"% Población ocupada con ingresos menores a 2SM:",'</strong>', round(df_municipio$PO2SM, 2),'</font>'
) %>% 
  lapply(htmltools::HTML)

saveRDS(state_popup_municipio, paste0(here::here(), "/2020/Output/PopUp/state_popup_municipio2.RDS"))
```


## Popup `Medio`

```{r}
## Paleta de colores                         
fillColor <- paleta[3]

df_municipio <- IMM_2020 %>% 
                 mutate(across(.cols = c(5:15,17), as.numeric)) %>%
                  filter(GM_2020 %in% "Medio")

## Etiquetas del mapa a nivel municipal 
state_popup_municipio <- paste('<font size="3"; font-family: "Century Gothic"><strong>',"Estado:",'</strong>','<strong>', df_municipio$NOM_ENT,'</strong></font>','<br/>',
                               '<font size="2.5"; font-family: "Century Gothic"><strong>',"Municipio:",'</strong>','<strong>', df_municipio$NOM_MUN,'</strong></font>','<br/>',
                               '<font size="2"; font-family: "Century Gothic"><strong>',"Población Total:",'</strong></font>','<font size="2"; font-family: "Century Gothic"; color=', fillColor,'><strong>', formatC(df_municipio$POB_TOT, format = "f", big.mark = " ", digits = 0),'</strong></font>','<br/>',
                               '<font size="2"; font-family: "Century Gothic"><strong>',"Grado de marginación:",'</strong></font>','<font size="2"; font-family: "Century Gothic"; color=', fillColor,'><strong>', df_municipio$GM_2020,'</strong></font>','<br/>',
                               '<font size="2"; font-family: "Century Gothic"><strong>',"Índice de marginación:",'</strong></font>','<font size="2"; font-family: "Century Gothic"; color=', fillColor,'><strong>', format(round(df_municipio$IM_2020, 2), nsmall = 1),'</strong></font>','<br/>',
                               '<font size="2"; font-family: "Century Gothic"><strong>',"Índice de marginación normalizado:",'</strong></font>','<font size="2"; font-family: "Century Gothic"; color=', fillColor,'><strong>', format(round(df_municipio$IMN_2020, 2), nsmall = 1),'</strong></font>','<br/>',
                               '<font size="1.5"; font-family: "Century Gothic"><strong>',"% Población de 15 años y más analfabeta:",'</strong>', round(df_municipio$ANALF, 2),'</font>','<br/>',
                               '<font size="1.5"; font-family: "Century Gothic"><strong>',"% Población de 15 años o más sin educación básica:",'</strong>', round(df_municipio$SBASC, 2),'</font>','<br/>',
                               '<font size="1.5"; font-family: "Century Gothic"><strong>',"% Ocupantes en viviendas paticulares sin drenaje ni excusado:",'</strong>', round(df_municipio$OVSDE, 2),'<br/>', 
                               '<font size="1.5"; font-family: "Century Gothic"><strong>',"% Ocupantes en viviendas paticulares sin energía eléctrica:",'</strong>', round(df_municipio$OVSEE, 2),'</font>','<br/>', 
                               '<font size="1.5"; font-family: "Century Gothic"><strong>',"% Ocupantes en viviendas paticulares sin agua entubada:",'</strong>', round(df_municipio$OVSAE, 2),'</font>','<br/>',
                               '<font size="1.5"; font-family: "Century Gothic"><strong>',"% Ocupantes en viviendas paticulares con piso de tierra:",'</strong>', round(df_municipio$OVPT, 2),'</font>','<br/>',
                               '<font size="1.5"; font-family: "Century Gothic"><strong>',"% Viviendas paticulares con algún nivel de hacinamiento:",'</strong>', round(df_municipio$VHAC, 2),'</font>','<br/>',
                               '<font size="1.5"; font-family: "Century Gothic"><strong>',"% Población en localidades con menos de \n 5 000 habitantes:",'</strong>', round(df_municipio$PL.5000, 2),'</font>','<br/>',
                               '<font size="1.5"; font-family: "Century Gothic"><strong>',"% Población ocupada con ingresos menores a 2SM:",'</strong>', round(df_municipio$PO2SM, 2),'</font>'
) %>% 
  lapply(htmltools::HTML)

saveRDS(state_popup_municipio, paste0(here::here(), "/2020/Output/PopUp/state_popup_municipio3.RDS"))
```


## Popup `Bajo`

```{r}
## Paleta de colores                         
fillColor <- paleta[4]

df_municipio <- IMM_2020 %>% 
                 mutate(across(.cols = c(5:15,17), as.numeric)) %>%
                  filter(GM_2020 %in% "Bajo")

## Etiquetas del mapa a nivel municipal 
state_popup_municipio <- paste('<font size="3"; font-family: "Century Gothic"><strong>',"Estado:",'</strong>','<strong>', df_municipio$NOM_ENT,'</strong></font>','<br/>',
                               '<font size="2.5"; font-family: "Century Gothic"><strong>',"Municipio:",'</strong>','<strong>', df_municipio$NOM_MUN,'</strong></font>','<br/>',
                               '<font size="2"; font-family: "Century Gothic"><strong>',"Población Total:",'</strong></font>','<font size="2"; font-family: "Century Gothic"; color=', fillColor,'><strong>', formatC(df_municipio$POB_TOT, format = "f", big.mark = " ", digits = 0),'</strong></font>','<br/>',
                               '<font size="2"; font-family: "Century Gothic"><strong>',"Grado de marginación:",'</strong></font>','<font size="2"; font-family: "Century Gothic"; color=', fillColor,'><strong>', df_municipio$GM_2020,'</strong></font>','<br/>',
                               '<font size="2"; font-family: "Century Gothic"><strong>',"Índice de marginación:",'</strong></font>','<font size="2"; font-family: "Century Gothic"; color=', fillColor,'><strong>', format(round(df_municipio$IM_2020, 2), nsmall = 1),'</strong></font>','<br/>',
                               '<font size="2"; font-family: "Century Gothic"><strong>',"Índice de marginación normalizado:",'</strong></font>','<font size="2"; font-family: "Century Gothic"; color=', fillColor,'><strong>', format(round(df_municipio$IMN_2020, 2), nsmall = 1),'</strong></font>','<br/>',
                               '<font size="1.5"; font-family: "Century Gothic"><strong>',"% Población de 15 años y más analfabeta:",'</strong>', round(df_municipio$ANALF, 2),'</font>','<br/>',
                               '<font size="1.5"; font-family: "Century Gothic"><strong>',"% Población de 15 años o más sin educación básica:",'</strong>', round(df_municipio$SBASC, 2),'</font>','<br/>',
                               '<font size="1.5"; font-family: "Century Gothic"><strong>',"% Ocupantes en viviendas paticulares sin drenaje ni excusado:",'</strong>', round(df_municipio$OVSDE, 2),'<br/>', 
                               '<font size="1.5"; font-family: "Century Gothic"><strong>',"% Ocupantes en viviendas paticulares sin energía eléctrica:",'</strong>', round(df_municipio$OVSEE, 2),'</font>','<br/>', 
                               '<font size="1.5"; font-family: "Century Gothic"><strong>',"% Ocupantes en viviendas paticulares sin agua entubada:",'</strong>', round(df_municipio$OVSAE, 2),'</font>','<br/>',
                               '<font size="1.5"; font-family: "Century Gothic"><strong>',"% Ocupantes en viviendas paticulares con piso de tierra:",'</strong>', round(df_municipio$OVPT, 2),'</font>','<br/>',
                               '<font size="1.5"; font-family: "Century Gothic"><strong>',"% Viviendas paticulares con algún nivel de hacinamiento:",'</strong>', round(df_municipio$VHAC, 2),'</font>','<br/>',
                               '<font size="1.5"; font-family: "Century Gothic"><strong>',"% Población en localidades con menos de \n 5 000 habitantes:",'</strong>', round(df_municipio$PL.5000, 2),'</font>','<br/>',
                               '<font size="1.5"; font-family: "Century Gothic"><strong>',"% Población ocupada con ingresos menores a 2SM:",'</strong>', round(df_municipio$PO2SM, 2),'</font>'
) %>% 
  lapply(htmltools::HTML)

saveRDS(state_popup_municipio, paste0(here::here(), "/2020/Output/PopUp/state_popup_municipio4.RDS"))
```

## Popup `Muy bajo`

```{r}
## Paleta de colores                         
fillColor <- paleta[5]

df_municipio <- IMM_2020 %>% 
                 mutate(across(.cols = c(5:15,17), as.numeric)) %>%
                  filter(GM_2020 %in% "Muy bajo")

## Etiquetas del mapa a nivel municipal 
state_popup_municipio <- paste('<font size="3"; font-family: "Century Gothic"><strong>',"Estado:",'</strong>','<strong>', df_municipio$NOM_ENT,'</strong></font>','<br/>',
                               '<font size="2.5"; font-family: "Century Gothic"><strong>',"Municipio:",'</strong>','<strong>', df_municipio$NOM_MUN,'</strong></font>','<br/>',
                               '<font size="2"; font-family: "Century Gothic"><strong>',"Población Total:",'</strong></font>','<font size="2"; font-family: "Century Gothic"; color=', fillColor,'><strong>', formatC(df_municipio$POB_TOT, format = "f", big.mark = " ", digits = 0),'</strong></font>','<br/>',
                               '<font size="2"; font-family: "Century Gothic"><strong>',"Grado de marginación:",'</strong></font>','<font size="2"; font-family: "Century Gothic"; color=', fillColor,'><strong>', df_municipio$GM_2020,'</strong></font>','<br/>',
                               '<font size="2"; font-family: "Century Gothic"><strong>',"Índice de marginación:",'</strong></font>','<font size="2"; font-family: "Century Gothic"; color=', fillColor,'><strong>', format(round(df_municipio$IM_2020, 2), nsmall = 1),'</strong></font>','<br/>',
                               '<font size="2"; font-family: "Century Gothic"><strong>',"Índice de marginación normalizado:",'</strong></font>','<font size="2"; font-family: "Century Gothic"; color=', fillColor,'><strong>', format(round(df_municipio$IMN_2020, 2), nsmall = 1),'</strong></font>','<br/>',
                               '<font size="1.5"; font-family: "Century Gothic"><strong>',"% Población de 15 años y más analfabeta:",'</strong>', round(df_municipio$ANALF, 2),'</font>','<br/>',
                               '<font size="1.5"; font-family: "Century Gothic"><strong>',"% Población de 15 años o más sin educación básica:",'</strong>', round(df_municipio$SBASC, 2),'</font>','<br/>',
                               '<font size="1.5"; font-family: "Century Gothic"><strong>',"% Ocupantes en viviendas paticulares sin drenaje ni excusado:",'</strong>', round(df_municipio$OVSDE, 2),'<br/>', 
                               '<font size="1.5"; font-family: "Century Gothic"><strong>',"% Ocupantes en viviendas paticulares sin energía eléctrica:",'</strong>', round(df_municipio$OVSEE, 2),'</font>','<br/>', 
                               '<font size="1.5"; font-family: "Century Gothic"><strong>',"% Ocupantes en viviendas paticulares sin agua entubada:",'</strong>', round(df_municipio$OVSAE, 2),'</font>','<br/>',
                               '<font size="1.5"; font-family: "Century Gothic"><strong>',"% Ocupantes en viviendas paticulares con piso de tierra:",'</strong>', round(df_municipio$OVPT, 2),'</font>','<br/>',
                               '<font size="1.5"; font-family: "Century Gothic"><strong>',"% Viviendas paticulares con algún nivel de hacinamiento:",'</strong>', round(df_municipio$VHAC, 2),'</font>','<br/>',
                               '<font size="1.5"; font-family: "Century Gothic"><strong>',"% Población en localidades con menos de \n 5 000 habitantes:",'</strong>', round(df_municipio$PL.5000, 2),'</font>','<br/>',
                               '<font size="1.5"; font-family: "Century Gothic"><strong>',"% Población ocupada con ingresos menores a 2SM:",'</strong>', round(df_municipio$PO2SM, 2),'</font>'
) %>% 
  lapply(htmltools::HTML)

saveRDS(state_popup_municipio, paste0(here::here(), "/2020/Output/PopUp/state_popup_municipio5.RDS"))
```
